# docs/swagger.yaml
basePath: /api
definitions:
...
  global.CredentialsLogin:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  global.PARTNER:
    properties:
      active:
        type: string
      address:
        type: string
      city:
        type: string
      comment:
        type: string
      country:
        type: string
      create_date:
        type: string
      create_user:
        type: integer
      email:
        type: string
      is_logged_in:
        type: string
      last_login_date:
        type: string
      last_logout_date:
        type: string
      npwp:
        type: string
      partner_id:
        type: integer
      partner_name:
        type: string
      partner_profile:
        type: string
      password:
        type: string
      phone_num:
        type: string
      state:
        type: string
      tax:
        type: string
      type_partner:
        type: string
      update_date:
        type: string
      update_user:
        type: integer
      username:
        type: string
      website_url:
        type: string
      zip_code:
        type: string
    type: object
...
  global.USER:
    properties:
      active:
        type: string
      comment:
        type: string
      create_date:
        type: string
      create_user:
        type: integer
      employee_id:
        type: integer
      is_logged_in:
        type: string
      last_login_date:
        type: string
      last_login_ip:
        type: string
      last_logout_date:
        type: string
      online_status:
        type: string
      partner_id:
        type: integer
      token:
        type: string
      update_date:
        type: string
      update_user:
        type: integer
      user_group_id:
        type: integer
      user_id:
        type: integer
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a Api Documentation for SIPIL 2022 with Golang Backend, Gin
    Gonic Framework, GORM MySQL, and JWT Authentication
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API DOCUMENTATION - API FOR SIPIL 2022
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login User
      operationId: Authentication
      parameters:
      - description: Input username & password
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/global.CredentialsLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Login User
      tags:
      - Users
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register User
      parameters:
      - description: Register User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/global.USER'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Register User
      tags:
      - Users
...
  /partner:
    get:
      consumes:
      - application/json
      description: API untuk mengambil semua partner yang terdapat dalam database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Mengambil Semua Partner
      tags:
      - Partner
    post:
      consumes:
      - application/json
      description: API untuk membuat partner baru.
      parameters:
      - description: Partner Data
        in: body
        name: partner
        required: true
        schema:
          $ref: '#/definitions/global.PARTNER'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Membuat partner baru.
      tags:
      - Partner
  /partner/{id}:
    delete:
      consumes:
      - application/json
      description: API untuk delete partner.
      parameters:
      - description: Module ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Delete partner.
      tags:
      - Partner
    get:
      consumes:
      - application/json
      description: API untuk mencari partner by ID yang terdapat dalam database.
      parameters:
      - description: Partner ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Mengambil Partner by ID
      tags:
      - Partner
    put:
      consumes:
      - application/json
      description: API untuk update partner.
      parameters:
      - description: Partner Data
        in: body
        name: partner
        required: true
        schema:
          $ref: '#/definitions/global.PARTNER'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Update partner.
      tags:
      - Partner
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"